Allow _SC_NPROCESSORS_CONF in sysconf(3) calls
gRPC has changed to call sysconf(_SC_NPROCESSORS_CONF), so we should
support it.

Also make sysconf return a long (as specified by POSIX) but use int64_t
in transit across the bridge.

Finally, rename a few variables in existing sysconf(3) code and change
the int_syscall_return field of SyscallsTestOutput from an int32 to an
int64 so that the sysconf test is guaranteed to not suffer from integer
overflow.

PiperOrigin-RevId: 206379523

+++ asylo/platform/arch/include/trusted/host_calls.h
@@ -189,7 +189,7 @@ int enc_untrusted_clock_gettime(clockid_t clk_id, struct timespec *tp);
 //////////////////////////////////////
 
 int enc_untrusted_pipe(int pipefd[2]);
-int enc_untrusted_sysconf(int name);
+int64_t enc_untrusted_sysconf(int name);
 uint32_t enc_untrusted_sleep(uint32_t seconds);
 int enc_untrusted_usleep(useconds_t usec);
 uid_t enc_untrusted_getuid();
